refactor: add package car
