+++ src/main/java/car/domain/Avante.java
@@ -1,11 +1,12 @@
-package domain;
+package car.domain;
 
-public class Sonata extends CarImpl {
-    private static final double DISTANCE_PER_LITER = 10;
-    private static final String NAME = "Sonata";
+public class Avante extends CarImpl {
+
+    private static final double DISTANCE_PER_LITER = 15;
+    private static final String NAME = "Avante";
     private final double tripDistance;
 
-    public Sonata (double tripDistance) {
+    public Avante(double tripDistance) {
         this.tripDistance = tripDistance;
     }
 
@@ -23,5 +24,4 @@ public class Sonata extends CarImpl {
     public String getName() {
         return NAME;
     }
-
 }
