+++ src/main/java/car/domain/Car.java
@@ -1,4 +1,4 @@
-package domain;
+package car.domain;
 
 public interface Car {
     double getChargeQuantity();
