+++ src/main/java/car/domain/RentCompany.java
@@ -1,4 +1,4 @@
-package domain;
+package car.domain;
 
 import java.util.ArrayList;
 import java.util.List;
