+++ src/test/java/car/domain/RentCompanyTest.java
@@ -1,4 +1,4 @@
-package domain;
+package car.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
